hm2_eth: fix 64-bit build error
Before this change, the following diagnostic is shown when building on 64-bit
platforms:
    warning: field precision specifier '.*' expects argument of type 'int',
        but argument 4 has type 'size_t' [-Wformat]
on 64 bit platforms where 'int' and 'size_t' are different sizes, it is likely
that this ends up passing incorrect arguments to vsprintf, which could lead to
a crash.

+++ src/hal/drivers/mesa-hostmot2/hm2_eth.c
@@ -404,7 +404,7 @@ static void hm2_eth_probe() {
 
     LL_PRINT("discovered %.*s\n", 16, board_name);
 
-    rtapi_snprintf(board->llio.name, sizeof(board->llio.name), "hm2_%.*s.%d", strlen(llio_name), llio_name, boards_count);
+    rtapi_snprintf(board->llio.name, sizeof(board->llio.name), "hm2_%.*s.%d", (int)strlen(llio_name), llio_name, boards_count);
 
     board->llio.comp_id = comp_id;
     board->llio.private = board;
